feat(cmd/push): parse requirements from plugins
When no requirements are provided by cli flags, the push
command will try to get them from the plugins. Plugins are
provided as shared objects, so falcoctl will load them if
are built for the same platform such the one where falcoctl
is running. When no requirements are provided and falcoctl
is not able to parse them from the plugin then it logs a
warning message and continues.

Signed-off-by: Aldo Lacuku <aldo@lacuku.eu>